Modified margin architecture.
The margins have been made way simpler. Instead of having elements working as
margins, the margin value of the screen is set so that the screen now sits in the
center of the screen. I guess I just made a complicated thing in the first place
because it felt like a cool solution.
This change came to be because some games that use a low scene resolution might have
gotten some artifacts where some pixel columns looked thicker than the others, or pixel
rows look shorter than others. This happened when the SCALE value was not an integer.
Now, the SCALE is set to an integer using a floor function, and this causes the scene
to not use the whole height (or width) of the screen. Now there are no problems.